refactor: Team 타입 중복 제거 및 commons로 통합
- teamSelectPage/types/teamSelect.ts에서 Team 타입 정의 제거
- commons/types/battle.ts의 Team 타입을 사용하도록 변경
- 영향받는 파일: index.tsx, TeamCard.tsx, Step4TeamSelect.tsx
- 모든 테스트 통과 확인 (122개)
